demo-service/requirements.txt
fastapi
uvicorn
requests

# OTel core
opentelemetry-distro
opentelemetry-exporter-otlp

# Instrumentations
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
